動的制御
