+++ server/package.json
@@ -34,6 +34,7 @@
         "react-modern-calendar-datepicker": "^3.1.6",
         "react-scripts": "^4.0.3",
         "react-three-fiber": "^6.0.13",
+        "react-timer-hook": "^3.0.5",
         "styled-components": "^5.3.3",
         "tailwindcss": "^3.0.2",
         "three": "^0.135.0",
